pyproject.toml
[tool.poetry]
name = "aisprayer"
version = "0.1.0"
description = "Detect unwanted animals in front of camera and sprays water on them"
authors = ["Christoffer Kongstad <christoffersk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
gpiozero = "~1.5"
picamera = "~1.13"
pillow = "~7.2"
requests = "~2.24"